+++ cistem/viewers/__init__.py
@@ -1,0 +1,5 @@
+from pyworkflow.utils import weakImport
+from viewers import CtffindViewer, ProtUnblurViewer
+
+with weakImport("tomo"):
+    from tomo_viewers import CtfEstimationTomoViewerCistem
