+++ cistem/viewers/viewers.py
@@ -26,20 +26,15 @@
 # *
 # **************************************************************************
 
-import os
-from matplotlib.figure import Figure
-
 from pyworkflow.protocol.params import LabelParam
 from pyworkflow.utils import removeExt, cleanPath
 from pyworkflow.viewer import DESKTOP_TKINTER, Viewer
 from pyworkflow.gui.project import ProjectWindow
 from pwem.viewers import CtfView, EmPlotter, MicrographsView, EmProtocolViewer
 import pwem.viewers.showj as showj
-from pwem.emlib.image import ImageHandler
 from pwem.objects import SetOfMovies
-from tomo.viewers.viewers_data import CtfEstimationTomoViewer
 
-from .protocols import CistemProtCTFFind, CistemProtUnblur, CistemProtTsCtffind
+from cistem.protocols import CistemProtCTFFind, CistemProtUnblur
 
 
 def createCtfPlot(ctfSet, ctfId):
@@ -229,44 +224,3 @@ class ProtUnblurViewer(EmProtocolViewer):
     def _findFailedMovies(self, item, row):
         if item.getObjId() not in self.failedList:
             setattr(item, "_appendItem", False)
-
-
-class CtfEstimationTomoViewerCistem(CtfEstimationTomoViewer):
-    """ This class implements a view using Tkinter CtfEstimationListDialog
-    and the CtfEstimationTreeProvider.
-    """
-    _targets = [CistemProtTsCtffind]
-
-    def plot1D(self, ctfSet, ctfId):
-        ctfModel = ctfSet[ctfId]
-        psdFn = ctfModel.getPsdFile()
-        fn = os.path.join(removeExt(psdFn) + '_avrot.txt')
-
-        xplotter = EmPlotter(windowTitle='CTFFind results')
-        plot_title = '%s # %d\n' % (ctfSet.getTsId(), ctfId) + getPlotSubtitle(ctfModel)
-        a = xplotter.createSubPlot(plot_title, 'Spacial frequency (1/A)',
-                                   'Amplitude (or cross-correlation)')
-        legendName = ['Amplitude spectrum',
-                      'CTF Fit',
-                      'Quality of fit']
-        res = _getValues(fn)
-        for y in ['amp', 'fit', 'quality']:
-            a.plot(res['freq'], res[y])
-        xplotter.showLegend(legendName, loc='upper right')
-        a.set_ylim([-0.1, 1.1])
-        a.grid(True)
-
-        return xplotter
-
-    def plot2D(self, ctfSet, ctfId):
-        ctfModel = ctfSet[ctfId]
-        psdFn = ctfModel.getPsdFile()
-        img = ImageHandler().read(psdFn)
-        fig = Figure(figsize=(7, 7), dpi=100)
-        psdPlot = fig.add_subplot(111)
-        psdPlot.get_xaxis().set_visible(False)
-        psdPlot.get_yaxis().set_visible(False)
-        psdPlot.set_title('%s # %d\n' % (ctfSet.getTsId(), ctfId) + getPlotSubtitle(ctfModel))
-        psdPlot.imshow(img.getData(), cmap='gray')
-
-        return fig
